+++ src/main/java/com/fredrikw/fructose/genetic/operators/UniformCrossover.java
@@ -9,10 +9,10 @@ public class UniformCrossover implements Crossover {
 	private float mixingRatio = 0.5F;
 	
 	@Override
-	public float[] crossover(float[] genesA, float[] genesB) {
+	public int[] crossover(int[] genesA, int[] genesB) {
 		Random random = ThreadLocalRandom.current();
 		int length = Math.min(genesA.length, genesB.length);
-		float[] result = new float[length];
+		int[] result = new int[length];
 		
 		for (int i=0; i<length; i++) {
 			if (random.nextFloat() < mixingRatio) {
